Remove guarding around adding direction
I wrote this game about eight years ago and I'd do everything very differently now. It's not clear that I'd want to do this defensive programming at this point in the code; in any case it's blocking getting any further because the key string returned has changed so now e.g. "Down" is "↓". I need to rewrite this code to use the KeyEvent virtual key codes, and this line was preventing the string even being moved on because of the `isDirection()` method.