Replace all tabs with spaces
Using `:retab`.

There are no other changes to these files.

I use vim for everything now. If I was working in Java full time then I would use an IDE but this is just for fun so might as well use the editor I like.

The fileformat is dos; I need to retab before changing it to mac otherwise it loses all line breaks.