Pass KeyEvent through to method
Rather than pulling out the keyCode, do this in the case statement in the method.

It's still worth having the extra method because the method name indicates what is being done on `keyPressed`, i.e. when the key is pressed, do X. The method name could be better though and I will change it in a subsequent commit.

This commit also moves a stray `}` - it's easier to see formatting after my fileformat changes.