Close window on game over
Before this change the game would finish and the JFrame would remain in place. In the absence of the options printed to the board noted in the comment, it's better just to close the board and restart it from the console if desired rather than abandoning the user.

That also gives the answer to the comment I've removed in this commit. I could still give keyboard input because the JFrame still existed.

Future developments could include printing an option to start a new game to the board, and so would remove the `dispose()` call.